# internal/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API request bodies and responses are usually JSON-encoded, except for binary data and some of the OAuth2 endpoints. Note that the `Content-Type` header must be set to `application/json` for this, as the request may otherwise fail with error 400.\nWhen clients have a valid access token, e.g. obtained through the `POST /api/v1/session` or `POST /api/v1/oauth/token` endpoint, they can use a standard Bearer Authorization header to authenticate their requests. Submitting the access token with a custom `X-Auth-Token` header is supported as well.",
        "title": "PhotoPrism API",
        "contact": {},
        "version": "v1"
    },
    "host": "demo.photoprism.app",
    "paths": {
...
        "/api/v1/geo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "finds photos and returns results as JSON, so they can be displayed on a map or in a viewer",
                "operationId": "SearchGeo",
                "parameters": [
                    {
                        "maximum": 100000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "maximum number of files",
                        "name": "count",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100000,
                        "minimum": 0,
                        "type": "integer",
                        "description": "file offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "excludes private pictures",
                        "name": "public",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5,
                            6,
                            7
                        ],
                        "type": "integer",
                        "description": "minimum quality score (1-7)",
                        "name": "quality",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "album uid",
                        "name": "s",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "photo path",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "is type video",
                        "name": "video",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/search.GeoResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "i18n.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "search.GeoResult": {
            "type": "object",
            "properties": {
                "Caption": {
                    "type": "string"
                },
                "Duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "Hash": {
                    "type": "string"
                },
                "Height": {
                    "type": "integer"
                },
                "Lat": {
                    "type": "number"
                },
                "Lng": {
                    "type": "number"
                },
                "TakenAt": {
                    "type": "string"
                },
                "TakenAtLocal": {
                    "type": "string"
                },
                "TimeZone": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "Width": {
                    "type": "integer"
                }
            }
        },
...
        "time.Duration": {
            "type": "integer",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000,
                1,
                1000,
                1000000,
                1000000000,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}